+++ sbin/dump/optr.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: optr.c,v 1.32 2013/04/16 19:25:59 deraadt Exp $	*/
+/*	$OpenBSD: optr.c,v 1.33 2013/04/25 06:43:20 otto Exp $	*/
 /*	$NetBSD: optr.c,v 1.11 1997/05/27 08:34:36 mrg Exp $	*/
 
 /*-
@@ -209,9 +209,10 @@ timeest(void)
 		deltat = tstart_writing - tnow +
 			(1.0 * (tnow - tstart_writing))
 			/ blockswritten * tapesize;
-		msg("%3.2f%% done, finished in %d:%02d\n",
-			(blockswritten * 100.0) / tapesize,
-			deltat / 3600, (deltat % 3600) / 60);
+		msg("%3.2f%% done, finished in %lld:%02lld\n",
+		    (blockswritten * 100.0) / tapesize,
+		    (long long)deltat / 3600,
+		    ((long long)deltat % 3600) / 60);
 	}
 }
